Rulesets now have their own file.
This change is primarily for future rulesets. If we just keep adding more classes to rulesets.js, it will become a monolithic file.
